Split mpl visualizer into different files
